Add "Hello, World!" build script and source
The intent of this build script is to demonstrate the JVM software
model's treatment of the simplest possible use case in Java programming:
compiling and running a main method that prints "Hello, World!" to the
console.

 - The main method and its enclosing class are self-explanatory.

 - The build script's `compile` task performs the equivalent of the
   following `javac` invocation:

    javac src/main/java/Main.java -d build/classes/

 - The build script's `run` task performs the equivalent of the
   following `java` invocation:

    java -cp build/classes/ Main
    Hello, World!

Most of the build script, including all task declarations will
eventually be moved into plugins, eventually leaving nothing more than a
simple, high-level declaration of an "application spec" or similar.
Subsequent commits to this build script should be subtractive in nature
as we support more and more of what's necessary "out of the box" in
Gradle core.

What's seen here has been transcribed with minor modifications from its
original location at [1].

[1]: https://github.com/gradle/features/blob/146abd95/core/building-components/component-spec-as-root/build.gradle