Rename mySources => sources, myClasses => classes
In the interest of keeping everything as simple as possible, and to see
if it's natural to use this naming. Note that prior to the previous
commit it was not possible to use the name 'sources' at the top level of
the model block, as it conflicted with the DefaultProjectSourceSet of
the same name contributed by LanguageBasePlugin. Now that the source set
in question is declared within the components block, this conflict is no
longer an issue.